(Bug #23) Redesign backends service

+++ src/main/java/net/rwx/maven/asciidoc/services/impl/FopServiceImpl.java
@@ -18,21 +18,19 @@ package net.rwx.maven.asciidoc.services.impl;
 
 import java.io.BufferedOutputStream;
 import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
-import java.io.IOException;
 import java.io.OutputStream;
 import javax.xml.transform.Result;
 import javax.xml.transform.Source;
 import javax.xml.transform.Transformer;
-import javax.xml.transform.TransformerConfigurationException;
-import javax.xml.transform.TransformerException;
 import javax.xml.transform.TransformerFactory;
 import javax.xml.transform.sax.SAXResult;
 import javax.xml.transform.stream.StreamSource;
+import net.rwx.maven.asciidoc.backends.Backend;
+import net.rwx.maven.asciidoc.configuration.Document;
+import net.rwx.maven.asciidoc.constants.Extension;
 import net.rwx.maven.asciidoc.services.FopService;
 import net.rwx.maven.asciidoc.utils.FileUtils;
-import org.apache.fop.apps.FOPException;
 import org.apache.fop.apps.Fop;
 import org.apache.fop.apps.FopFactory;
 import org.apache.fop.apps.MimeConstants;
@@ -44,33 +42,45 @@ import org.apache.fop.apps.MimeConstants;
 public class FopServiceImpl extends RootServiceImpl implements FopService {
     
     @Override
-    public void execute( String input, String output ) throws FileNotFoundException, FOPException, TransformerConfigurationException, TransformerException, IOException {
-        OutputStream out = null;
-        try {
-            
-            FopFactory fopFactory = FopFactory.newInstance();
+    public void execute( String inputPath, Document document, Backend backend ) throws Exception {
+        setOuputPath(inputPath, backend);
+        if( backend.isIsFopTransformation() ) {
+            OutputStream out = null;
+            try {
 
-            File outputFile = new File( output );
-            if( outputFile.exists() ) {
-                FileUtils.forceDelete( outputFile );
+                FopFactory fopFactory = FopFactory.newInstance();
+
+                File outputFile = new File( getOuputPath() );
+                if( outputFile.exists() ) {
+                    FileUtils.forceDelete( outputFile );
+                }
+                out = new BufferedOutputStream( new FileOutputStream( outputFile) );
+
+                Fop fop = fopFactory.newFop( MimeConstants.MIME_PDF, out );
+
+                TransformerFactory factory = TransformerFactory.newInstance(); 
+                Transformer transformer = factory.newTransformer(); 
+
+                File inputFile = new File( inputPath );
+                FileUtils.copyFileToDirectory( inputFile, new File("/tmp") );
+                Source src = new StreamSource( inputFile );
+                Result res = new SAXResult( fop.getDefaultHandler() );
+
+                transformer.transform( src, res );
+            } finally { 
+                if( out != null ) {
+                    out.close();
+                }
             }
-            out = new BufferedOutputStream( new FileOutputStream( outputFile) );
+        }
+    }
 
-            Fop fop = fopFactory.newFop( MimeConstants.MIME_PDF, out );
-
-            TransformerFactory factory = TransformerFactory.newInstance(); 
-            Transformer transformer = factory.newTransformer(); 
-
-            File inputFile = new File( input );
-            FileUtils.copyFileToDirectory( inputFile, new File("/tmp") );
-            Source src = new StreamSource( inputFile );
-            Result res = new SAXResult( fop.getDefaultHandler() );
-
-            transformer.transform( src, res );
-        } finally { 
-            if( out != null ) {
-                out.close();
-            }
+    @Override
+    protected void setOuputPath(String inputPath, Backend backend) {
+        if( backend.isIsFopTransformation() ) {
+            setOutputPath(inputPath, Extension.PDF);
+        }else {
+            setOutputPath(inputPath);
         }
     }
 }
